Add code to spot another kind of broen PDF file
If the highest number in an xref (calculated from the initial value
plus the number of value) exceeds the /Size defined in the associated
trailer dictionary then that's wrong, but we can probably work around
it if that's the only problem, so just issue a warning.

Add an error for yesterday's fix for the problem of using an object
defined in the xref table as free. That's an error because in general
we can't fix it (the test file for that bug works by sheer fluke)