Update MSVC makefiles.
Firstly, another version of nmake to cope with. In fixing this, I
note that we haven't been setting MS_TOOLSET_VERSION in some cases.
For building from the IDE this evidently doesn't matter, but I suspect
it is required for building using nmake. Add those back in.

Thanks also to Julian for spotting that the logic to cope with
"Community" vs "Professional" editions was failing due to extra
quotes. Fixed here.

Also, msvclib.mak hadn't been updated for VS2017 or VS2019. Copy
the relevant sections across. Possibly we should pull this code out
to a separate file that we can !include so we only have to maintain
it in one place.