Fix error handling leading to counting down an object twice.
We transfer the reference to a derefrenced object from 'Next' to
'outline', then process outline before retrieving a new 'Next'.

However if the processing of outline returned an error we would exit and
cleanup, counting down both outline and Next, which would at that point
be referencing the same object but only having counted it up once.

Fix by setting Next to NULL when we transfer the reference.