pdfwrite - fix memory leak when converting names into references
When writing certain kinds of metadata from pdfmarks we need to convert
a named object into an indirect reference. The code for doing this
assumed that the gs_param_string array was arranged so that each
gs_param_string's data member pointed to an individually allocated
chunk of memory, so it was safe to replace it.

This had 2 problems; firstly the original data buffer was not freed,
leading to memory leaks. Secondly, and more seriously, not all
gs_param_string_arrays are organised this way. Sometimes the array
has a single data pointer containing all the entries sequentially,
each member of the array is a gs_param_string where the data member
points into a location in the string. We cannot free and reallocate
the data members from that kind of string array.

SO this commit modifies the code to make a new copy of the data in the
param_string_array in the format expected by the remaining code, and
frees the copied data afterwards. This fixes the meory leak and avoids
the potential crashes caused by the incorrect assumption.

Note that the assumption in the code is true when the input is from the
PostScript interpreter, but is not true for the input from pdfi and
obviously potentially any other front-end.