pdfwrite - Limit XUID to 16 values and don't emit it for subset fonts
Bug 703225 "The document breaks after repeated processing"

As noted by Peter Cherepanov, we currently preserve the XUID in the font
when writing it out from pdfwrite, and the PDF interpreter extends any
existing XUID (or adds a new one) every time we process the file.

The PLRM documents an Adobe limitation of 16 entries in an XUID array
and it appears Acrobat has the same limitation.

This commit adds new flags to the type 1 and type 2 font writers so that
we only write an XUID when we are *not* subsetting the font (as the use
of an XUID on two different subset fonts would be incorrect). It also
only emits the first 16 values from an XUID even if we aren't subsetting
the font.