Fix bug 702786: Ignore null node in Page tree
Thanks to Peter Cherepanov for this analysis and the patch,

The sample file has an element in the /Pages array that cannot be resolved,
and following the PDF spec is interpreted as null.

This patch makes Ghostscript ignore null nodes in the page tree. In turn,
this causes a missing page error, which was considered a fatal error before.

-dPDFSTOPONERROR causes Ghostscript to not proceed with this broken file.

Regression testing shows no differences with other files.