PDF Interpreter - Clean up stack after Form Xobject for transparency Group
Bug #703217 "Error: /typecheck in /--pdfshowpage_finish--"

This is basically the patch supplied by Peter Cherepanov, which uses
the existing mechanism for restoring the stack to a count stored
before the execution of the Form XObject.

However, Peter's patch caused a regression with the file for bug 697655
when run with pdfwrite, because that file consumes more operands from the
stack than it should, leading to us trying to pop -1 elements.

This commit only pops a positive number of objects, and has a slightly
different error text.