pdfwrite - fix memory leak on error path
When writing outlines via a pdfmark, if the code failed for any reason
then we would leak various objects. As usual this is cleaned up for us
by the garbage collector when running in PostScript but leaks when we
aren't using the garbage collector.

There are likely to be many, many more paths that leak like this, the
only way to find them is going to be a complete audit of the pdfwrite
source, which would be a major undertaking. For now we'll fix them as
we find them.