Don't assume that pdfi_read_token returned a token, just because there was no error return.
If we try to read bytes beyond the EOF then pdfi_read_token returns 0
but does not create an object on the stack. Generally we check the stack
to make sure we have an object before we try to use it but
pdfi_read_object() wasn't doing that. Add some checking code at each
stage to ensure we have a token.