Fix a memory leak and handle undefined ref better
Fix memory leak on error path when building a string had an error.

Also fixed the cause of the error, which was trying to build a string
ref for an undefined reference.
This was with the file tests_private/pdf/forms/v1.6/july+205++202007.pdf,
which is a broken file.
Now just build a "N 0 R" ref string for it even though the object doesn't
exist.  There might be a better way to fix this, not sure.

Also fixed a random unused variable warning.