Reintroduce and update cff std strings "encoding" PS file
and move it to "lib" since we don't actually use it in Postscript any more.

It's only used if we ever need to regenerate the C representation of "known
encodings".

Also, update to include missing strings.

Finally, fix the tool that relies on this "encoding" - .namestring is no
longer exposed as a Postscript operator, so define a local equivalent. And
update the example invocation in the comments