Fix conversion of PostScript caluclatr functions
The code to convert type 4 (PostScript calculator) functions into
Ghostscript's internal binary representation of a type 4 function had
an error when dealing with ifelse constructs.

The internal representation starts an if with an 'if' code, then
two bytes denoting the length of the following stream which contains the
converted 'if' clause. An ifelse is similar, an if value, two byte
length then an else value and two byte length of the else clause.

However..... in the case of an 'ifelse' the length for the 'if' must
include the three byte sequence for the header of the else clause.

This is tricky for us to do because, unlike the PostScript code which
has the entire function as an executable array and can peek ahead to
see if this is an if or ifelse, we only have what we hace already
parsed.

So retain a pointer to the 'if' and patch the value afterwards if we
find that we actually had an 'ifelse'. This could be done more neatly
but this works.