PDF interpreter - improve commit for bug #703217
In the event that the execution of a Form XObject for a transparency
Group consumed more operands than were pushed, we left the number
of operands on the stack. This just cleans up the stack.

The file for bug #697655, with the pdfwrite device, is the only known
case of this, and the file fails with pdfwrite anyway, so this is not a
truly vital fix, but its best to tidy up.