Fax Decode filter - add a simple sanity check
I have a test file (source unknown) e249869cf6a250726711c21fda42763c_signal_sigsegv_a418b9_5694_4827.pdf
which has a number of type 3 bitmap glyphs with invalid CCITTFaxDecode
parameters. The (optional) Columns key in the dictionary has an
associated value of 2^32-1 in one case and an even larger value in
another.

The new pdfi interpreter causes a seg fault in the CCITTFaxDecoder
because this breaks a 32-bit signed integer on Windows, causing the
'raster' value to be negative, when added to the memory pointer this
causes us to try and read from before the beginning of the data buffer.

Ghostscript's PDF interpreter doesn't trigger this because it fails
much earlier and doesn't try to execute these glyphs but I'm certain
it would be possible to modify the file so that it did.

Adding a simple negative check and throwing an error if it is, resolves
the problem.