PDF interpreter - handle indirect objects in Encrypt dictionary
Bug #703272 "This PDF files converts to tiff using the oitsample program, but the font turns from English to what looks like chinese"

The Encrypt dictionary and the subsidiary StdCF dictionary are stored
as indirect objects (presumably in order to make the file larger). The
decryption code wasn't expecting that.

Add code to dereference the objects. While we're here, add code to check
the types of the objects and ensure they are correct, and throw an
error if they are not.