PDF interpreter - check type of Group entries in Form XObjects
Bug #703271 "Using oitsample to convert some pdf files to tif errors out with "The export process timed out.""

This is for the larger (205 pages) of the two files. The file is badly
broken, 3 fonts point to objects which are not dictionaries but are
simply integers. In addition at least one of the Form XObjects has a
Group entry which also points to an object which is an integer, not a
dictionary.

The Group was corrupting the operand stack leading to later errors. This
commit checks the type of the Group (and the BBox since we're making
changes anyway) and simply aborts the Group if the type is incorrect.

Obviously this is going to lead to incorrect output, but at least we
can carry on to later pages.

Note that Acrobat throws errors on pages 168, 169 and 171-175 of this
file and renders blank pages for them, so the file is clearly very badly
broken.