Fixes for high level forms/modAP
1) If there is no valid AP, then just delete the AP.
This fixes Bug697951.pdf which has an AP that is an empty dictionary.

2) Fixes for file 'ccs digital.pdf' which is from Ken
This sample has the same annotation referenced from multiple pages, and it
was doing wonky things and getting the wrong stream.

First, this change requires that we make sure the AP doesn't get dereferenced
(which in the normal case would turn it into a STREAM object).
That requires a change to pdfi_check_annot_for_transparency() as well
as pdfi_annot_preserve_modAP().

Second, we need to make sure to leave the indirect ref in place,
because it might be referenced in multiple places.  So instead of
creating a new ref, we just modify the old one.

So now there are two extra variables in an indirect ref that are only
used for high level forms.  These are the form object number, and a
flag indicating it's a high level form.  Those get setup in modAP and
then cleared down in pdfi_obj_indirect_str() after they are used.

This means each AP for an annotation will get rendered as a separate
high level form, even if there are multiple instances referencing the
same AP.  That probably takes unnecessary extra space in the case that
it is reused on the same page, but for the case where they are on
separate pages, this is actually necessary in case the pages have
different graphics states.