Use specified Box, honour -dFitPage and /Rotate when run from PostScript
Previously we always used the MediaBox to set the media size, we did not
apply any /Rotate value form the PDF file and did not respect FitPage.
We also ignored UserUnit.

To enable these the PageInfo dictionary returned from .PDFPageinfo has
been extended to include all the Boxes from the PDF page dictionary as
well as any /Rotate value and any UserUnit.

The media setup as well as the initial CTM now considers whether to use
the current media size or the specified Box (eg -dUseCropBox) and
whether that should be the media size or used to scale the content on to
the current media. We also use non-zero llx and lly Box values to shift
the content, to cater for clipping, particularly important with CropBox.

I've done my best to test this, but I suspect there will be some
combinations which don't work as expected, we'll need to fix those as
they come up. However as far as I can tell the results are much
improved. A couple of files which are missing text are rotated
incorrectly because of AutoRotatePages.