Remove paths stored in --permit-file* when sanitising.
Bug #703338 "Hide permitted paths from invocation comment"

As noted in the bug report, the --permit-file* control arguments are
not replaced with '?' when sanitising the arguments. This leads to an
information leak with the 'Invocation' comment emitted by pdfwrite.

This commit adds checking for the -- switches and sanitises the
permit-file* cases.