Tifflib fix: Thunder RLE can fail to decode last run.
GPDL decode of tests_private/tiff/text.tif gives valgrind errors
due to the "thunder decode" failing to extract the last run of
bytes.

The logic in the decoder, presumably intended to spot overruns of
data is incorrect, in that runs that end at the end of a row
(npixels == maxpixels) will not be decoded.

Fix this by limiting 'n', the number of pixels to copy. Note that
npixels is updated by the 'unlimited' value to ensure the error
reporting at the end of the loop still works.