Fixes/changes to default subclass device procedures.
The 'output_page' should pull the PageCount from the child device (that
is the only real fix.)

The other changes are to get rid of all of the return statements that
cannot be reached. The 'else' is retained as a comment /* else */ in
case someone can't figure that out. It is sort of surprisng that we
don't get 'statement cannot be reached' messages out of any compiler or
static analysis tool.