Fix issues with windows C# demos/csharp/api/ghostnet.cs
Make sure to explicitly tell C# that the callback delegates will use a Cdecl
calling convention, which means that the caller (in this case the ghostscript)
will clean the stack.   Also some minor changes to the project settings.
The default for debug is to use gpdldllxx.dll  while the release versions will
use gsddllxx.dll just to help with the testing.  Finally, make sure to *always*
copy the dll from the appropriate gs directory, even when the C# application
did not need to do a build (as it was already built)