Remove pdfi_array_peek
The function pdfi_array_peek, along with the requirement to be able to
not optionally not resolve indirect objects and optionally not store
resolved indirect objects in the array in place of the indirect object
meant that pdfi_array_fetch() required that resolved indirect objects
had a reference count of at least 2.

Because we generally add newly resolved objects to the object cache this
was generally true, but this was never intended to be a requirement. In
addition, if an object could not be resolved we returned a null object
with a reference count of only 1, which was then counted down and freed
before being used, leading to memory corruption.

I can't see a way to fix pdfi_array_peek() and still cater for objects
not having a reference count of 2, so this commit removes the function
and replaces all uses of it with code which expects to handle the
reference counting. That is, it expects the object it is returned to
have it's reference count incremented to reflect the reference the
caller is holding, and to count it down when it is finished with it.

For some reason which I have not investigated, the refactoring of the
pdfi_get_glyph_metrics() function has apparently changed the logic
slightly leading to a progression in a file with Arabic text.

This commit resolves the seg faults I was seeing with pdfi running
under Ghostscript and does not leak any memory in my usual testing.