Implement strip_tile_rect_devn for pdf14 device
The default strip_tile_rect_devn method returns an error
and has a comment that the method is only implemented
for devices that support devn, which the comment says
are only the memory planar devices.  However, the pdf14 device
also supports devn colors and has to be able to handle
pattern tiles that have spot colors.  When such files
are encountered, the interpreter will just report an error in
the file and continue on its way if we are in non-clist mode.
When in clist mode, the error will cause ghostscript to stop
execution.

This implementation borrows heavily from gx_default_strip_tile_rectangle
to deal with all the logic of where we are on the tile,
it then borrows heavily from pdf14_copy_mono but
using devn colors and dealing with the case of
one of the colors possibly being gx_no_color_index.