Add malloc_.h to gp_ntfs.c
Bug #703310 "gp_ntfs.c uses malloc but does not include "malloc_.h""

There doesn't seem to be a problem with adding malloc_.h to the C file
it matches what we do with other files, doesn't produce any warnings
and passes regression testing.

Using WIN32_LEAN_AND_MEAN however causes the compiler to fail with a
number of errors, so we won't be adding that. The nature of the errors
look like these would be difficult to solve and the benefit of this
preprocessor directive is moot, even Microsoft insiders describe it
as 'largely useless'.