Alter PDF arrays to initially be fully populated with null objects instead of NULL pointers.
This is primarily for safety, it means we can get an object from
an array index and be guaranteed to get a valid PDF object. Clients
can then interrogate hte type of the object (which they probably
need to do anyway) instead of checking first for a NULL pointer.

The array_from_stack function doesn't do this. There must already
be enough objects on the stack to fill the array (because the size
is determined from the number of objects on the stack) and they must
be valid objects, so there's no point in slowing it down.

Added a pdfi_dit_alloc() to match the pdfi_array_alloc, we don't
populate this initially as you can't retrieve from an index with
a dictionary so we don't need to do that. This is really just present
currently to be symmetrical with arrays. However we may need it in
future if we ever change the operation of dictionaries, so it
makes sense to implement it now.

Check to see that keys are not NULL when retrieving them. It should
not be possible for them to be NULL, because we compact dictionaries
when we delete from them and only retrieve up to 'entries' rather
than 'size'. Again this is future-proofing.