Guard against indirect /Length in an XRef stream
Bug #703372 "Unable to find Trailer"

The file has an XRef stream (a compressed cross-reference), where the
/Length of the stream is itself an indirect object. We can't dereference
the Length, because to do so we would need to look up its object number
in the cross-reference!

Oddly the PDF specification does not state that the /Length must be a
direct object.

This commit wraps the code which retrieves the /Length in a stopped and
if it fails we fall back to the assumption that the length is invalid
and apply our normal recovery strategy.