More jpeg passthrough error case fixes.
Ensure that on an error return, we communicate the end of JPEG passthrough to
the (potential) recipient(s) of the data.

Basically, we need to ensure that JPEG passthrough is completed before the
last time we return from s_DCTD_process for the current stream - i.e. whether
it is an EOD condition, or an error condition.