oss-fuzz 27399: Handle NaN and Inf in object type conversions
In the core of cvs detect unrepresentable floating point numbers and throw
an error (note: not using fpclassify or similar because they are C99 additions).

If we hit such a case, throw an error, in this case undefinedresult - we
cannot use rangecheck because that gets special handling which also triggers
the uninitialized value error.