Better fix for psdcmyk indeterminisms.
When calculating buffer sizes, we now take the presence/absence of
tags into account. This must be copied into the temporary device
structure from the target.

Simply setting it to 0 will stop the indeterminisms, but will give
problems with potential future devices (psdcmyktags?).

Accordingly, I've removed the memset, and copied the value, meaning
we can rely on valgrind to help spot such problems in future.