Allow pdfi to use an external name table
When running under Ghostscript, and using the pdfwrite device, the final
PDF file is not written out until the pdfwrite device is closed. This
can (generally will be) be after the PDF interpreter has been terminated
and freed.

The problem is that the pdfwrite device retains copies of the fonts it
needs (which were defined during the course of the PDF interpretation)
and those copies use name indices which were assigned by the PDF
interpreter and cannot be resolved back to names after the interpreter
has exited.

The only real way to solve this is for the PDF interpreter to use the
enclosing PostScript name table (and methods for turning names into
indices and back to names), which will persist until the PostScript
interpreter is closed, by which time the pdfwrite device will already
have terminated.

To do this we add a couple of function declarations to the PDF context
which point to the name table functions for turning a name into an index
and an index into a name. Intially we assign those to the pdfi functions
and the .PDFInit PostScript operator replaces them with functions
appropraite to it.

This commit also cleans up some of the pdfi interpreter startup and
closedown to ensure we don't free the name table too early, and to
make certain that if we are using the pdfwrite device it is restored
away before we exit the interpreter (this does not happen when running
under Ghostscript as the device is already present and protected by
other levels of gsave). This ensures that the PDF file is written out
before we close down the pdfi interpreter.

This solves the seg faults we were seeing with pdfwrite when running
pdfi inside Ghostscript. Unfortunately it also adds some new errors.

Investigation of the errors shows that our current infrastructure
around names; the enumerate_glyph() functions and pdfi_cnecode_char()
are inadequate. Addressing this is going to be a longer term project
so I'm choosing to add the code to fix the crash problems and accept
the new errors for now.