txtwrite - Address some memory handling problems
Bug #703143 "segfault in ps2txt with certain large file"

There are two problems here; firstly the textw_text_release()
function is called *from* gs_text_release, and should not be calling
that function. This was the initial cause of the seg fault.

Secondly the txtwrite custom text enumerator was not declaring the
'pte_fallback' text enumerator member, which meant that it could be
relocated without updating the txtwrite text enumerator, leading to
seg faults further through the processing.

This resolves the seg faults on the supplied test file for me, but there
are still memory problems (memory leaks in fact) revealed by running
under Memento. I'll address these in a later commit or commits.