PDF interpreter - work around broken ObjStms (again)
Bug #703448 " Ghostscript can't read files that poppler, mupdf and Firefox and others can read"

The problem is in an OObjStm, a compressed representation of various
PDF objects. In the example file the ObjStm returns a lone 'mark'
object for one of the compressed objects, which is not legal.

This appears to be the font T1_2 on page 87.

The mark confuses our counting of the returned objects and that leads
to the interpreter falling in a heap.

There's no trivial way to address this, so I've chosen to use the
approach used for error handling in other places; push a specific name
onto the stack, and then use that instead of a mark to delimit the
portion of the stack of interest.

Obviously nothing is going to rescue the broken font; but since this is
a PDF file which has been OCR'ed I don't think this is a problem as we
will not to try to use it for rendering as the text is all 'drawn' in
text rendering mode 3 (neither fill nor stroke).

Because we keep using this I've also added a couple of utility procedures
CountToKey and ClearToKey, which work similarly to countomark and
cleartomark. Because I wanted to be able to detect a missing key on the
stack CountToKey returns a boolean, not just a number.

I've also used that approach around the verify_page_tree call so that
we don't end up with a pile of junk on the stack if it should fall
over similar errors in future.

Finally I tested with the customer supplied file which originally
inspired the changes in resolveobjectstream to detect the last kind of
broken ObjStm and that file continues to work.