Replace -dSimulateOverprint boolean with multi-valued -dOverprint
Internally the name is changed to "overprint_control" which has enum values
of gs_overprint_control_t. This is in preparation for Michael Vrhel's
overprint simulation implementation.

Documentation of the change and the new parameter specifics added to the
doc/Use.htm file. Also added change to gs_init.ps so that command line use
of -dSimulateOverprint=true maps to -dOverprint=/enable and similarly,
-dSimulateOverprint=false maps to -dOverprint=/disable with a warning.