Bug 703454: Handle multiple CharString definitions
The test file for this has a number of Type 1 fonts whose CharStrings
dictionary definitions contain more than one entry for several glyph names.
According to Postscript rules (which Type 1 fonts are supposed to follow),
defining a key/value pair for which the key already exists in the dictionary
should overwrite the earlier value - i.e. the later definition takes
precedence.

It appears that the non-Postscript based implementations used in other
PDF consumers do not honor that, and the earlier definition ends up taking
precedence.

To work around various other PDF embedded Type 1 font breakages, we push an
extra dictionary onto the dict stack which contains definitions for missing
procedures, and various other "fixups".

To work around this problem, we add a new definition of the Postscript "def"
operator which, *only* for the CharStrings dictionary, will not allow existing
keys to have their values replaced.