Preserve unread bytes when layering SubFileDecode filter
This came about as a result of bug #703448. That file has an ObjStm
where the object numbers are immediately followed by the first object
(object 0) with no white space (this is totally reasonable).

The problem is that when reading the tokens we can read ahead by up to
one byte, which we then 'unread'. We do that by storing the byte in the
(pdfi) stream object's unget_buffer. However, when reading the actual
objects, we start by layering a SubFileDecode filter onto the stream
which is reading and decompressing the ObjStm, and that simply ignores
any unread bytes!

Fixed here by having pdfi_apply_SubFileDecode_filter() check to see if
the underlying stream has any unread bytes. If it does we copy them to
the new (pdfi) stream object and set the underlying (pdfi) stream
'unread size' to 0. We also reduce the EODCount by that number of bytes,
if EODCount is not 0.

I'm surprised this hasn't come up before.