oss-fuzz 25846: Store "printable" refs for errors
When we get an op/dict/exec stack overflow error, we take a copy of the stack
in question, and store it in an array (or as much of it as we feasibly can).
That allows us to clear enough stack space to allow the error handler(s) to run
without losing the information needed to report the error.

With the exec stack, however, that can (and usually does) include references to
"internal" objects, which are really just pointers to arbitrary C structures,
(t_struct and t_astruct ref types). Those can contain pointers referencing
objects of unrelated and uncertain lifespans, so persistent storage of them is
potentially problematic.

So, when we create that array representation, replace every t_struct and
t_astruct object with a printable string representation.

This isn't a big problem for speed because a) exec stack overflow is a rare
error b) for exec stack overflow, we already have to traverse the objects
checking for another condition, and c) we already do (basically) the same
conversion considerably later, for the benefit of the error handler.