Remove a (now) unnecessary function call
Since the cleanup of function type 0 data streams was added to the
graphics library, we no longer need to do it in the PDF interpreter.

I've left the function in place for now, in case we discover anything
else that wants to use it, and until I have a chance to do memory
leak testing.