oss-fuzz 22842: sanity check the claimed number of objects in a PDF
This test case is a file that we identify as a broken PDF, and attempt to
repair.

As part of trying to repair the xref, we an extremely large value integer which
we take to mean the number of objects in the file, and then attempt to loop
over them. The result is, in practice, an endless loop.

On the basis that the very barest PDF object:
1 0 obj
endobj

takes 15 bytes to define, we add a check: if the number of claimed objects is
greater than the number of bytes in the file divided by 15, then the number
is not to be trusted, and we dicard it.