pdfwrite - further checks on Title for PDF/A
Bug #703486 "PDF containing just a JPEG converted to PDF/A-1b fails rule 6.7.3-1"

The PDF file, produced by ImageMagick 7 has a /Title in the Document
Information dictionary which appears to be UTF-16BE encoded, but is
lacking the BOM. It also has a 2-byte NULL appended which seems likely
to be an error as well (PDF strings are not NULL-terminated).

This is valid (if not useful) as a string with PDFDocEncoding, though
Acrobat displays an empty string in the Document Properties, but when
converted to a text string for the output Document Information
dictionary, and a UTF-8 string for the XMP, the two strings are not
byte-for-byte identical, resulting in a PDF/A validation error.

Add some more checking for character codes which will result in UTF-8
and PDFDocEncoding strings which do not match (anything outside the
range 0x20 to 0x7F, or escaped).