pdfwrite - don't try to linearize when the output file is not seekabkle
No bug report. The linearization (-dFastWebView) 'feature' of pdfwrite
relies upon being able to seek in the output file. If we can't do that
then we can't linearize the file.

If linearization is enabled, check the output file is seekable; if it
is not then warn the user and abort linearization.