Rework the PostScript/PDF interface slightly
While writing documentation it became apparent that the opaque object
we pass around with the PDF interpreter context in it should not, now
be called a PDFfile object, because that's not what it is, and it's
confusing since we no longer use .PDFfile to create it.

I've renamed it to a PDFcontext. I've also reworked the .PDFInit
operator slightly so that it does not crash in vmreclaim after an
error in .PDFInit, and to sort out the stack. Previously if we passed
in a dictionary argument it would remain on the stack.