Change the misnamed 'depth' value from 1 to 0.
No differences expected as this falls into either less than 0 or 0
and greater. The comment in gstrans.h about this being the 'transparency
stack depth' is utterly wrong, and the parameter name is misleading
but it's what we have.

Add a comment regarding the simulate overprint feature which is currently
not implemented.