Bug 702910: Fix mkromfs for THREADSAFE build
Since local versions of errprintf_nomem() and  errprintf() are mostly
identical, we can just call errprintf() with a dummy mem argument.

The "mem" argument for errprintf() inside GhostPDL is used to ensure
the error prints are directed to/through the correct gs_lib_ctx
instance for the current thread

This is safe because, mkromfs is only using during the build, and won't
ever implement threading, thus has no need to worry about the extra
parameter.

(slightly tweaked patch to keep working without THREADSAFE)