Don't entirely flush the stack on error in pdfi_dereference
There are conditions under which we can continue after a dereferencing
error, and at least one of them (/Length in a stream object which is
the /Contents for a page) we must keep the actual stream object on the
stack.

Exhibited by file 0688736_02252_HP LaserJet 3100-3150 All-in-One Service Manual.pdf
from the bugzilla PDF collection.