Change macro names to be less confusing.
The graphics state contains pairs of color/colorspaces. The 0th ones
are known as 'current' ones. The 1st ones have historically been known
as 'alt' ones.

This nomenclature is held to be confusing, as 'alternate'
colors/colorspaces mean something quite different. We therefore update
the macros etc to refer to these as 'swapped' ones, which fits better
into our usage.