Set object pointer to NULL after ownership transfer
Ownership of the object pointed to by 'Next' was transferred to 'child'
and we expect 'Next' to be reused (a few lines above). However, if an
error occurred in pdfi_dict_get_no_store_R() then we would end up with
child and Next having the same value, and on exit we would count down
both pointers, causing us to free the object while the cache was still
pointing at it.

This caused memory corruption when we later tried to evict the object
from the cache, corrupting a totally different object, which then caused
a seg fault when we tried to free the Pages tree on exit.

The file causing an error was one of the Bugzilla file collection
0689936_04181_PDF_For_Artifex.pdf